Added a scratchwork folder for one-off documents
This is basically for everything that’s not an exercise but calls for
more space than a note in iBooks. (I’m making highlights and taking
notes as I go—they’re just not part of this repository because they’re
more for my own reference than anything else, and because I suspect
they’re stored in a relatively opaque format that wouldn’t play nice
with git.) The first thing I’m adding here is a document wherein I work
through an example of the counting change algorithm of 1.2.2.